backend/pyproject.toml
[project]
name = "miraat-backend"
version = "0.1.0"
description = "The backend for the Miraat application."
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.1",
    "python-dotenv==1.0.1",
    "SQLAlchemy==2.0.36",
    "psycopg2-binary==2.9.9",
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "Jinja2==3.1.4",
    "google-generativeai==0.8.3",
    "sentence-transformers==3.3.1",
    "scikit-learn==1.6.0",
    "torch==2.5.1",
    "transformers==4.46.3",
    "langchain==0.3.9",
    "python-multipart==0.0.20"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
]

[tool.setuptools]
packages = ["core_logic"]